pyproject.toml
[tool.poetry]
name = "datargs"
version = "0.1.0"
description = "Declerative, type-safe command line argument parsers from dataclasses and attrs classes"
authors = ["Roee Nizan <roeen30@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
dataclasses = {version = "^0.6", python = "3.6"}
attrs = {version = "^20.2.0", optional = true }

[tool.poetry.extras]
attrs = ["attrs"]

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.0.2"
attrs = "^20.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
